chore(docs): Optimized array indexing in the command bindings example
# Project State

## Current Focus
Optimized array indexing in the command bindings example

## Context
The change removes an unnecessary cast operation in the command bindings example, improving code clarity and potentially small performance gains.

## Completed
- [x] Removed redundant `as usize` cast in array index calculation

## In Progress
- [x] No active work in progress

## Blockers
- None

## Next Steps
1. Verify the change doesn't affect functionality
2. Consider similar optimizations in other examples
---
category: chore
scope: docs
---

+++ examples/_cookbook/command_bindings.rs
@@ -339,7 +339,7 @@ impl Widget for CommandBindings {
             }
         }
         for x in 1..w - 1 {
-            let idx = ((h - 1) * w + x) as usize;
+            let idx = (h - 1) * w + x;
             p.cells[idx].bg = t.surface_elevated;
         }
